move server-http and httplib to its own file
